checked for both lines are parrllel to x-axis
What changes you have made?
added if both lines are parallel to x-axis it should return true for
isParallelTo

Why this change is necessary?
should validate when both lines are parrallel to x-axis

What is the impact of the change?
it is validating when both lines are parrallel to x-axis