hasPoint method is added to the line class
What changes you have made?
added hasPoint method to the line class

Why this change is necessary?
to check whether given point is present in the given line or not

What is the impact of the change?
it will check the given point is present in the line or not