modified hasPoint method in the Line class
What changes you have made?
modified hasPoint method in line class

Why this change is necessary?
to check point is present on the line if line is parallel to x or y axis

What is the impact of the change?
it is also checking point when line is parallel to x or y axis