hasPoint method is added to Rectangle class
What changes you have made?
hasPoint method is added to rectangle class

Why this change is necessary?
to find whether a point is present in present on the sides of rectangle
or not

What is the impact of the change?
it can determine whether the point is present on rectangle or not