updated isEqualTo method in Rectangle class
What changes you have made?
added condition when another diagnol of rectangle is given

Why this change is necessary?
to check two rectangles are equal when another diagnol of rectangle is given

What is the impact of the change?
it can check two rectangles are equal when another diagnol of rectangle
is given